Reformatted sql file
I want you guys to follow this format STRICTLY. Understand that this file is read by our program and every query is executed individually.

We cannot drop any triggers. Its okay even if we do not use procedures we will use fucntions which are even better.

No blank lines anywhere
-- comments shoudl start and end like this --
changing the delimiter is a seperate query not a combied one so keep it seperate.

I dont care what you write in the comments, without comments the parsing will fail, so you may fill it up with shit, but only one line